// carta-controller-deploy/carta-controller-v2/src/types.ts
export interface CartaExternalAuthConfig {
    issuers: string[];
    publicKeyLocation: string;
    keyAlgorithm: string;
    // Unique field to be used as username
    uniqueField: string;
    // User lookup table as text file in format <authenticated username> <system user>. If no user lookup is needed, leave this blank
    userLookupTable?: string;
    // Routes for refreshing access tokens and logging out
    tokenRefreshAddress: string;
    logoutAddress: string;
}
// ...
export type TokenPayload = {
    username?: string;
    scripting?: boolean;
    iss?: string;
    refresh?: boolean;
    expires_in?: number;
};

// Token verifier function
export type Verifier = (cookieString: string) => TokenPayload | Promise<TokenPayload> | undefined;
// Map for looking up system user name from authenticated user name
export type UserMap = Map<string, string>;